+++ src/libawkward/Identities.cpp
@@ -141,8 +141,8 @@ namespace awkward {
   void IdentitiesOf<T>::nbytes_part(std::map<size_t, int64_t>& largest) const {
     size_t x = (size_t)ptr_.get();
     auto it = largest.find(x);
-    if (it == largest.end()  ||  it->second < sizeof(T)*length_*width_) {
-      largest[x] = sizeof(T)*length_*width_;
+    if (it == largest.end()  ||  it->second < (int64_t)(sizeof(T)*length_*width_)) {
+      largest[x] = (int64_t)(sizeof(T)*length_*width_);
     }
   }
